requirements.txt
Django==4.2.7
djangorestframework==3.14.0
python-dateutil==2.8.2
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
selenium==4.15.2
black==23.11.0
flake8==6.1.0
gunicorn==21.2.0
coverage==7.3.2
sphinx==7.2.6